refactor: update authService import paths for consistency across API routes
