+++ src/app/api/profile/password/route.ts
@@ -4,7 +4,7 @@ import { NextRequest, NextResponse } from 'next/server';
 import { authOptions } from '@/lib/authOptions';
 import { logError } from '@/lib/logger';
 
-import { authService } from '@/services';
+import { authService } from '@/services/auth/authService';
 
 import { ChangePasswordSchema } from '@/shared/validation/profileSchemas';
